feat: 기존 terms 폴더 구조 컨벤션에 맞게 변경
