+++ apps/web/src/widgets/auth/ui/index.ts
@@ -10,7 +10,10 @@ export { EnrollmentVerificationHeader } from './enrollment-verification-header';
 export { EnrollmentVerificationForm } from './enrollment-verification-form';
 export * from './enrollment-step-card';
 export * from './enrollment-step-card-header';
-export * from './terms';
+export { TermsBottomSheet } from './terms-bottom-sheet';
+export { TermsDetailContent, type TermsDetailContentTerm } from './terms-detail-content';
+export { TermsDetailDialog } from './terms-detail-dialog';
+export { TermsDialog } from './terms-dialog';
 export { SessionKeepConfirmModal } from './session-keep-confirm-modal';
 export { SignInImageSection } from './sign-in-image-section';
 export { OauthAdditionalInfoForm } from './oauth-additional-info-form';
