+++ apps/web/src/widgets/auth/ui/terms-bottom-sheet/index.ts
