+++ apps/web/src/widgets/auth/ui/terms-content/index.ts
@@ -1,0 +1,1 @@
+export { TermsContent } from './TermsContent';
