+++ apps/web/src/widgets/auth/ui/terms-dialog/index.ts
