--- apps/web/src/widgets/auth/ui/terms/index.ts
@@ -1,2 +1,0 @@
-export * from './bottom-sheet';
-export * from './dialog';
